api/requirements.txt
gunicorn
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
simpleaudio
six==1.15.0
tweepy==3.9.0
urllib3==1.26.2
Werkzeug==1.0.1
